XSPEC: support etable table models
Unfortunately there's no way to tell from a table-model file
whether it's multipicative (mtable) or exponential (etable), so
out-of-bnd information is required (the user has to explicitly
say what type is being used). The default assumes mtable.

Note that there is no support added to load_table_model, as the
support for XSPEC table models there has been deprecated for a
long time.