Tests: add basic WSTAT plot checks
What data is shown when we plot the data and model with stat=WSTAT?

We test both low-level and the UI layer just to check the same
information is being used. This does lead to repeated tests, which
is not ideal.

At the moment the WSTAT statistic does not "correct for" the model
signal in the data. The intention is that we will be changing this
in the next commit.

There are also tests of the backscal handling and what error or
warning messages are created.