Address review comments from @hamogu
Bump the copyright date as we never made it in 2020.

There was a discussion about the number of decimal places created
by the use of format, but it turns out that for this function it
isn't relevant since our tests are only used for integer values.
There are other "tostatement" versions in this module where the
comment makes more sense (i.e. this was probably just originally
from a copy-n-paste).