Docs: improve cache discussion and documentation
The main change is to add some information to the "Caching a model"
page, but there are several releated tweaks to docstrings.

The behavior of the cache system is hard to describe, so this doesn't
attempt to do so - the main aim is to document what is happening -
and we can hopefully clean the behavior up soon.