Rework cache counting
Move the counts out of the cache into a separate dictionary (it
could be a tuple as it might be faster, but not worth such
optimisations at this time).

We also now track all accesses, even those not being cached. This
can help point out why a cache stops being used because you've
done a fit and it has decided to turn off the cache. Why it does
this I don't know.