Add cache_clear and cache_status methods to models
The routines are modelled after functools.lru_cache, except that
that uses cache_info which returns a tuple and here we print to
stdout the cache status. This difference is in part because we
can have multiple components, each with a different cache, that
would need to be combined in some manner for a routine like
cache_info. It is much easier just to print out the response for
each component.