Tests: add explicit source plot tests for PHA data
As sherpa.astro.plot.SourcePlot does not use the "normal" routines
for evaluating the data to plot (that is, using the to_plot method
of the data object) we need to check the behavior for a number of
cases like

  - changing units
  - changing rate
  - changing the factor setting

The test has also changed from using integer values for comparison
back to an approximate equality test due to changes in how the data
is compared (e.g. multiplying by the bin width, particularly for
the wavelength check).