Hide AstroPy 4.2.1 FITS-related warnings
astropy.io.fits.open now create warnig messages about invalid FITS
structures when given a non-FITS file. Unfortunately this leads to
a lot of confusing messages for the user of some of the top-level
routines - like sherpa.astro.io.load_data - when given an ASCII
file, since the code will try to load a FITS file multiple times
before trying asn ASCII file.

We avoid this by hiding all warnings from astropy.io.fits when
we try to open a FITS file. This means that we will not see useful
warnings that are generated when the input is a FITS file, but I
think that is a valid compromise. We could re-create the warnings,
as we can change to something like

    with warning.catch_warnings() as ws:
        out = fits.open(fname)
        for w in ws:
            <somehow recreate the w>

but it's not completely obvious how to recreate the warnings (as
AstroPy has over-ridden the warnings module and I did not feel
it's worth the effort to understand what's going on here).