Fix small error in notice/ignore when no data is loaded
The notice and ignore functions were not error-ing out when
there was no data loaded because of two issues:

 - presumably ude to changes in the Python 2 to 3 transition
   a check for an empty dictionary was written in such a way that
   it never actually identified an empty dictionary
 - the handling of string filters was done before the "no data"
   check, leading to different code paths

The handling of the ignore function has been simplified, removing
a special case, as it just repeats code in notice. This simplifies
the code, and makes it easier to see what is going on, at the small
expense of making ignore do more in some cases. As this is not a
"hot path" for analysis the code simplification is preferred.