Simplify ignore_id
As with notice/ignore, make sure that ignore_id is as thin a wrapper
around notice_id as possible, to ensure we have consistent behavior
(e.g. checking for ids being valid if lo is a string, although we
already have an explicit check of at least some of these cases).