Change how completely "invalid" filters are handled (PHA)
The idea is that

    pha.notice(a, b)

where a and b lie either below or above the valid range for the
dataset are now ignored. Previously they would lead to the first
or last group to be noticed [or ignored, if ignore was called].

Although this improves some things (e.g. see the tests that were
added to check thes very behavior) it is not yet clear whether
this is a worthwhile change.

It is worthwhile noting that none of our existing tests (other
than those explicitly testing this behavior) have changed, which
is nice to see.

This commit includes several assert statements which are there to
document conditions I believe should hold. Ideally we would either
error out "nicely" if they don't hold, deal with them, or we can
just assume they hold, but I wanted to see if we have any cases of
the assertions failing before doing anything. The assert checks
should be removed or replaced before this enters a release.