Tests: remove xfail from notice/ignore tests
These tests are just to check that the code gives the same answer
as before, not that the result is sensible. This will make it
easier to see when the filtering has changed the result (before
we could have an xfail test have the result change but we'd be
none the wiser as long as it still didn't give the expected
answer, particularly when some of the expected answers were not
necessarily meaningful).