Tests: add internal checks of PHA filtering
This adds tests of internal routines (PHA _to_channel, which depends
on the unit setting) that are implicitly tested in some of the recent
tests added or updated (e.g. test_notice_channel_grouping) but I
wanted explicit tests as there appeared to be some iteresting behavior,
which is shown here in the wavelength filtering: a negative value
is treated differently to the energy case (even given the swap in
low/high behavior we see here). The idea is to add tests of the
existing behavior so we note when it has been changed.