Explicit error out if notice/ignore range is a string
The ui layer lets you use a string for the lower-bound of a range
but this is not supported in the data layer. There are explicit checks
for the standard data classes (a test is added for this) and so
we add an explicit check for the DataPHA class (needed since it
doesn't use the same filter code as the other classes), which leads
to a small cleanup in the code, and a test added.