Tests: check numpy conversion for counts/channels of DataPHA
We have some checks to make sure the user knows that input arguments
to the Data objects may be replaced when they are not NumPy arrays
(originally added in #614 I think). This doesn't quite work for the
DataPHA case since the tests are on the independent and dependent
axes (`x` and `y`) whereas the user has given channel and counts
(even though they are essentialy aliased together, even if not
actually aliased).

Before I touch this I wanted checks of what happens when the
channel or counts argument was not a NumPy array.