DataPHA constructor checks channels,counts arrays
There is an existing check for the dependent axis to make sure
it is a NumPy array, but it works oddly for the DataPHA case,
as we have a copy of the dependent axis as the counts field and
this isn't covered by the check. There are also places where we
assume the independent axis is a NumPy array (normally in the
line 'self.channel + 1').

We now explicitly check that both the counts and channel arrays
are NumPy arrays.

It is not clear what happens if the counts array has a mask - I
wiould expect things would get confused but I have not tested
this. Similarly we do not explicitly tell the user that any mask
on the independent axis is ignored.