Fix save_data when the output file exists
As @wmclaugh pointed out in my previous attempt to fix #1071 (in
PR #1066) the actual problem was a lot simpler - the optional
arguments were not being set correctly. This commit addresses this
(fixing the problematic line and other lines which were okay but
could show this problem if their callnig convention was changed).

The tests added in #1067 have had their xfail labels removed.