This is a doc-string only update for load_multi_XXX
THe previous example implied that a specturm may contain both positive
and negative orders in the same spectrum. That can happen only after
coadding - and in that case one would co-add the arfs as well.

fixes #982

Since this is a doctring-only update, I'll take the opportunity to see
if I can avoid wasting electrons for CI.

[ci skip]