BUGFIX: typo in finfo calls
The code had a closing bracket in the wrong place, as

  pow(numpy.float_(numpy.finfo(numpy.float32)).eps, 1.0 / 3.0)

is invalid. I assume the float_ was meant to coerce the finfo(..).eps
call, but as this is unnescessary we can simplify to

  pow(finfo(..).eps, 1.0 / 3.0)

Note that this bug has existed for a long time (the last change to
these lines was for 4.8.0 and the bug existed before this change).
It appears that this is only triggered in an "exceptional" case -
that is, handling a special case, but I we don't have *any*
tests of the NumDeriv-derived classes.

There is also one non-bug change, replacing float(...) by ... as
there is no need for the explicit conversion.