Tests: support use of data directory being local, not absolute
A test was checking for the start of a path to a file assuming
it was absolute, but it's not always true. As the path part of the
message check is not important remove the check for the /
character.

This is only needed for someone who says

    sherpa_test -D ../path/sherpa-test-data/sherpatest

If they use an absolute path it is okay.