Docs: improve sherpa.models._modelfcts.integrate1d
Add the optional arguments for this function, where I am
somewhat-guessing about their meaning (although logger has been
shown to be correct and errflag is a value that gets sent through
to the code but never used).

It is not expected that users will use this help so it is not
extensive.