Fix crash from grouping data using an unsupported method
The following requires the C++11 standard, but there is no
change in documentation to record this, or any attempt to
ensure this restriction in the setup.

I had not realised that the grouping support was limited to a
hard-coded set of routines, and found this out by crashing Python.
This fixes this, by explicitly checking that the group function
is supported and if it isn't raising a ValueError. Documentation
for do_group is added.