CI: remove coverage reporting for sherpa_test
Remove an addition added in #1001. The problem is that sherpa_test
is explicitly run outside the Sherpa git repository, and so there
is no wat to identify the commit for the coverage code, causing
problems both in running the coverage and then when sending the
coverage data to our coverage system. It is not-at-all obvious
how to work around this, so I'm just throwing in the towl.

This does mean we lose some coverage of less-used code - i.e.
fall through cases when the fits or plotting backends are not
available - but so be it.