Remove section on "bulding docs on travis" from the docs
Summary
This section is outdated (we don't use Travis any longer) and, in fact,
factually wrong. Docs are build in RTD for each PR and can be manually
inspected. However, keeping a seciton like this updated is a pain and
we see the RTD check mark on every PR, so I'm just removing it.

[skip ci]