Remove unnecessary code
When the "make_groups" option was being used we called do_group
with this option and then threw away the results and manually
created the groups. This is wasteful.

I checked to ensure that we had tests that hit this case and to
make sure that do_group was returning the same data, so it's
safe to remove.

I will note there is one subtle difference: do_group returns an
array of c, c+1, c+2, ... where c is the first channel number,
whereas the manual code always set c to 1. This means we could
see a difference in behavior if we ever have a PHA file which
doesn't start at channel 1, which is possible. We just do not
have any data like this in our tests.