Track the requested PHA filter range
Track the requested ranges applied to a PHA data set so that they
can be re-created when the grouping is turned on or off. The idea
is that the _filter_add method is used to store each request -
that is the low, high, units, and ignore flag - and then we
can use the _filter_recrate method to re-run all the steps. So
if you have done

    pha.notice(0.5, 7)
    pha.ignore(1.1, 1.2)
    pha.group_counts(15)
    pha.group_counts(20)

then it's as if you've done

    pha.group_counts(20)
    pha.notice(0.5, 7)
    pha.ignore(1.1, 1.2)

This is not all the store since there are three plot tests that
fail from sherpa/astro/ui/tests/test_astro_ui_plot.py

    test_pha1_get_model_plot_filtered
    test_pha1_model_plot_recalc[model-extraargs0-ModelHistogram-644-252]
    test_pha1_model_plot_recalc[model_component-extraargs1-ComponentModelPlot-644-252]

but they wil get fixed in a later commit (so that the test does not
need to be changed). There are also grouping changes that need to
be done to switch over to this scheme.