fix #1024
The PHA histogram plot, used for the model plot of PHA data, is
"special" in that it shows the ungrouped data even if the PHA
data is grouped. This requires the plot object to temporarily
ungroup the data, and in doing so there is a need to re-apply
the filter state. Due to changes in how the filter state works,
we actually have to change the behavior here - that is make
sure we use a channel range that matches the grouped data and
not just apply the filter to the ungrouped data, which would end
up selecting a smaller channel range than we want.

The upshot is that the handling of the prepare method of the
sherpa.astro.plot.ModelHistogram class has been made more
complicated. However, in doing so we have ended up fixing
issue #1024.