Ensure that grouping changes propagate (fix #1160)
The dynamic grouping routines call the group method, which means
that the filter will be re-applied to use the new scheme. This
means that the support for re-creating the filter range can be
removed here. Due to the subtle differences in the grouping
schemes we do get slightly different results. For the PHA model
histogram plot we decided to use the existing channel range, so
as to better-match the pre/post-grouped data, but here I believe
we want to apply the original filters using the new grouping
scheme, which can result in "edges" of the filters changing here.

The grouping attribute of the DataPHA class is now gated by a
property accessor which lets us enforce the constraint that the
groupnig data must be the right size and values, as well as
being to re-create the filter if the grouping is changed (when
it wasn't before, which is issue #1160).