Tests: add basic OrderPlot tests and fix an error-handling issue
Provide some very-basic tests of OrderPlot and, in doing so, fix
a check for an unusual situation. The code was using orders and
colors when it should have been using self.orders and self.colors,
but because there were no checks written for this class it was
never noticed before.