Style changes for OrderPlot
I do not understand why OrderPlot does what it does - in particular
dropping the first color and the choice for the orders setting
in prepare, but this is outside the filtering/grouping changes
so leave as is. The problem is that there are no tests to show
what this code is supposed to be doing.