Docs: improve top-level documentation for data
This adds some background informatiom for the sherpa.data and
sherpa.astro.data modules.

Note that this takes advantage of sphinx/RST support to provide
links in the RTD version of these pages, as the expense of making
the raw docstring - e.g. as viewed by help(sherpa.astro.data) -
to appear like Perl has crawled into our files (e.g. saying
:py:class:`DataPHA` rather than `DataPHA`). I am against adding
this sort of mark-up to those docstrings that users are going to
come across, but it's probably okay for these module-level
docstrings.