RTD point out that filtering features depend on the data
This is just to point out that there's more to life than the
simple example. Exactly where this extra functionality is to
discussed (both for the user and the developer) is for another time.

The niternal links have been moved from BaseData to Data to reflect
changes made to Sherpa a long-time ago but weren't being flagged
by Sphinx (i.e. this makes the text lniks rather than just labels).