Docs: document an internal routine
We have _get_ebins and _get_indep which are both mysterious and
now have *some* documentation, but it's still not really clear
what their intent is.

I added a TODO note that there is **very-limited** testing of
the multi-response case (a number of routines I added to cover
this sort of case but they are not using "real-world" data).