Comment change to clarify grouped setting
In reviewing the code I was worried that a situation like

    pha.group()
    pha.group_counts(20)
    pha.group_counts(30)

would not lead to a _filter_recreate call because we don't do
anything in _set_grouped if the flag is unchanged. However, this
is issue #1160 and I had forgotten that I had explicitly changed
the grouping attribute so that when it is changed we call
_filter_recreate, so we don't need to here. Hopefully the
somewhat-shorter comment added in this commit is enough to explain
this.