Add a comment related to issue #1166 - quality handling
It is likely that we have a number of places where we are not
handling quality data properly when the data is ungrouped. This
is just the one I noticed.

For example, pha has all quality values set to 2:

   >>> (pha.quality == 2).all()
   True

so that when grouped we have

   >>> pha.grouped
   True
   >>> pha.ignore_bad()
   >>> pha._get_ebins(group=True)
   (array([], dtype=float64), array([], dtype=float64))

but if you turn off grouping - either setting pha.grouped = False
or the flag, then you get all the channels:

   >>> a, b = pha._get_ebins(group=False)
   >>> a.size
   1024