Tests: add explicit UI image_xxx tests (requires DS9/XPA)
There are existing tests of the DS9 connection but at a low level,
not from the UI layer - e.g. no test of image_fit(). This adds
some basic tests.

Note that these tests can not reliably be combined with pytest-xdist,
so

    pytest -n auto

may well fail on these tests. This is because we only create a
single DS9 instance which is then connected to by all the tests,
resulting in potential corruption if two DS9 tests are run at the
same time (the state is undefined in this case). We could look at
supporting a per-instance DS9 connection, but that is a large
change. For now, if you have DS9 and XPA installed then I suggest
you have to run the tests serially.

Note that this is also true of the existing DS9 tests, it's just
that no-one has really run into this combination before (and I
have not checked the existing tests to see what they are actually
checking; it is possible they are very-simple and would not fail).