Move the image classes to a dictionary
Follow plot and contour types and store the image objects in a
dictionary rather than a bunch of attributes of the session class.
More importantly, the code has been updated so that the logic for
setting the image object (i.e. calling the prepare_image method
with the correct logic) has been moved to the get_xxx_image
routines since

a) there was a bug (now fixed) when image_model_component and
get_model_component_image were usng different logic - now we go
through a single method so we are correct by construction

b) it is easier to over-ride (evidenced by plot changes), even if
we do not make use of it here.