Ensure the method attribute is set to a callable.
Not many people will be changing the method attribute of the
ModelDomainRegridder1D class, but now we can now at least make
sure the argument is callable. I decided not to add further
checks to make sure that it accepts the correct arguments because
this is low-level code and, at the end of the day, we can't make
a perfect check. I also use a TypeError rather than a
Sherpa-specific error class as, again, this is low-level code,
and we have other routines that raise similar errors.

The test added in the previous check now passes.