Docs: add top-level docs for sherpa.models.parameter
Just describe the basics of parameter handling, as we just assume
this knowledge in the sherpa code base (there is discsusion of this
in the RTD user-level docs, but not here).

There is some interesting issue I have found with the link support
in writing this documentation, but I don't want to hold back the
documentation for a possible code investigation.