Tests: replace use of SherpaTestCase in parameter tests
This moves the SherpaTestCase cases to pytest-style. It fixes a
hidden error in one of the tests - a test called setattr incorectly,
raising a TypeError, but that is what the test was looking for (but
a different error). This case actually raises a ValueError, so I
changed the test.

The old code created TestParBase/Val classes in the SherpaTestCase
object, but moving these to the top level causes pytest to think
there should be tests present and there aren't. So the Test prefix
has been removed here. I plan to completely remove these classes in
the next commit as they don't add anything, but that's a change too
far for this commit.