Tests: add a couple of tests of link handling
In writing the parameter documentation I'd noticed a subtle
difference between setting .val or .link to a link expression,
so I started this PR to address this. Of course, I've now found
out I was not doing what I thought I was doing, but let's not
worry and ensure we have a simple test to check this.