Add has_pha_response check
The idea is to provide a single place where we can codify what
"containing a response" means. It is possible that more classes
should be added to the check (and it would be nicer to be able
to determine from the class itself), but they are issues for
later.

This is not used in the code, but will be.