model component PHA plots now auto-include the response (fix #1020)
Model-component plots - so generated by either plot_model_component
or get_model_component_plot - will now automatically include the
instrument response if one is not included manually. This makes
use of the new sherpa.astro.instrument.has_pha_response functionality.

Note that the documentation updates here are a bit wonky: the
plot_model_component changes are actually made to the sherpa.ui.utils
version because we don't need to overload it here. I plan to come
back to the long-standing question about the need for bkg_id support
for the source and model component plots, in which case we will need
to overload plot_model_component and fix the odd versionchanged issue.
I just don't feel it's worthwhile trying to fix in any other way
here.