Tests: explicit check of plot('model_component')
In reviewing PR #1022 which this work is based on I noted that
I had seen problems with plot('model_component'), so this adds
checks of this (and plot_model_component for good measure).
Fortunately whatever problem I saw in #1022 isn't shown here,
probably nbecause of changes to the plot code since that time
(or I'd made an error in thinking there was an error; a 50/50
chance I'd say).