Change PHA filter/grouping to send the name not function
Since do_group uses the name of the function let's make this the
actual argument to apply_grouping and hence apply_filter. As this
is technically a breaking change we still support sending in the
function (as before), to support anyone with old code, but this
now raises a deprecation warning.