Deprecate PHA "grouping" methods
With the change to apply_grouping we can remove the methods
which were only used for their name, not code. I would like to
remove them but am leaving them in for now in case users are
using them, but labelling them as deprecated.

We do not ever run these methods so I have not added any
tests for the deprecation.