RTD: note the apply_filter change for PHA files
There are two changes: the data/pha.rst and evaluation/examples.rst

The data version adds a section on 'Manipulating data' to (briefly)
describe the apply_filter and apply_grouping methods.

The evaluation version has been updated to use the new string
option when calling apply_filter and apply_grouping. It also
updates the example to use the PHA-specific plot classes
(DataPHAPlot, ModelHistogram).

A version-changed note for 4.14.0 has been added for the
apply_filter and apply_grouping methods of DataPHA.