Tests: Use pytest functionality in test_root
Rewrite the teststo use pytest functinoality (parameterizing a
test with arguments rather than having multiple functions that
only vary in three arguments). It also means we get greater
visibility in the case of a failure since we know that all
methods will be tried and not just a failure after one mrthod
for a particular problem.

There are two tests that are known to fail - previously we just
"ignored" them but now they are explicitly marked as a skip,
in case we ever decide to change the demuller2 code.