Tests: partial removal of SherpaTstCase from test_template
This changes the easier-to-convert use of SherpaTestCase.

Note that the test is not actually that useful - all it does
is check we can evaluate the model but no check on the returned
values.

Two commented-out tests have been removed (I tried running them
but it's not clear the functionality they were intended to test
was implemented).