Tests: remove SherpaTestCase from test_astro
A number of tests followed the pattern

  - run thread a
  - check
  - run thread a + '_ncpus'
  - check

I've broken these out so that each test just runs one thread and
it's parameterised to select either the "basic" or "+ ncpus"
version. Technically this is no-longer checking the same thing,
but none of the tests rely on running from an "unclean" session
so I think this change is beneficial (as you get better tracking
of potential errors). These tests could be further cleaned up,
as the structure is the same in most of them, but that is left
for the future as this change is extensive enough as is.