Tests: ensure FITS headers are copied in image
A recent PR means we now copy over the FITS header values for
images (#1155) so add a test to ensure this remains true. This
also does some random checks on the FITS properties of the file
and shows up some differences between crated and pyfits backends.

Note that issue #1206 describes some of these differences.