Improve OGIP header for PHA output files.
This does not address all issues: some of these should
proably be handled by the DataPHA class itself.

There are a number of keywords that we set which should be
identified from the DataPHA object (e.g. CHANNEL range and
whether the data is a HDUCLAS3=RATE rather than COUNT).

There is no attempt to support TYPE:II output (but we don't
really provide that capability).

Note that we can now read back in the PHA file we create
with write_pha when using crates.