Tests: basic get_filter checks for Data1D, Data1DInt, DataPHA
Although the PHA get_filter calls have existing checks, I wanted
a set of focussed tests as I look to address #1129. In doing so
I realised it would make sense to add some checks of Data1D and
Data1DInt (in particular the Data1DInt case as it is similar to
DataPHA but handled differently, and doesn't have issues with
grouping and a change in the independent axis units).

Of course, in doing so I've found a Data1DInt problem which
these tests exercise (and are marked as xfail).