Data1DInt filtering now reports start/end ranges
The response to

  d = Data1DInt('x', [1, 2, 4, 5], [2, 3, 5, 6], [0, 0, 0, 0])
  d.get_filter()

was to report the center of each bin as the filter range, so

  '1.5000:5.5000'

We now report the actual end points, so in this case

  '1.0000:6.0000'

For the Data1DInt case it probably doesn't make much difference,
as it's easy to map the filter description to the independent
axis. It is more of an issue for the DataPHA case but I'm tackling
the easier version first.