Tests: add a get_filter test
This doesn't really add anything new to the existing tests but it
was useful during development.

I also fix the numeric precision of the output of get_filter in
test_energy_filter_roundtrip to avoid changes as Python/NumPy
evolves.

The
sherpa/astro/tests/test_astro_data.py::test_pha_filter_simple_energy0
test is just to report the current behavior; it does not make sense,
and is something to do with channel starting at 0 - see #1040