Tests: add explicit tests of PHA/Data1DInt notice/ignore handling
The idea is to test the handling of a filter that matches the
upper edge of the response. The current output is not great
but we treat this as a regression test (and will hopefully
be able to change this).

We repear for the Data1DInt case as we have the same issue
there, but as the filtering is handled differently the results
don't quite agree.

See #1211 for more information.