Non-channel PHA filter ranges must be >= 0
It is now an error to use invalid values when filtering PHA
data - that is, when energy or wavelength units are in use
the notice and ignore low and high values must be >= 0.
Technically we should restrict to > 0 but using 0 is a
useful short-cut (a user could also use None but that's
is less user-friendly).

We also check that lo <= hi in all cases.