_get_ebins now errors out if all channels are bad
We had some code that would call _get_ebins and then recognize
that all channels had been masked out, causing DataError('nomask')
to be raised, and some that didn't. I have moved this error to
_get_ebins as it makes sense to have it there.

Since this only happens when the data is grouped we only have to
worry about this when group=True.

The behavior of _get_ebins with grouped data or not is a bit
odd, but that's a larger change than it's worth making here.