Channel PHA filters must be integers
It is not clear what to do when a non-integer value is used
for a channel filter in a notice() or ignore() call, so we
now raise a DataErr (similar to how the ranges are restricted
to >= 0 for non-channel ranges).

Alternatives include

  - x -> int(x)
  - x -> int(x + 0.5)  (i.e. nearest int)

We could support int(x), and I believe this is what the code
used to do, but I prefer the explicit check as channels are
discrete (there are some subtle potential differences with how
we handle channel filters for non-integer values if we improve
the filtering for energy/wavelength ranges).

We allow channel ranges to be negative since the OGIP standard
does not obviously exclude this.

Previous commits have changed a number of tests so that they
are not affected by this change (by adding a response to the
PHA dataset so filters can remain non-integers).