Data1DInt/DataPHA filtering to better handle start/end filters
- NOTE

This can change fit results as the number of bins included
in the fit for DataPHA or Data1DInt datasets can be changed
with this commit. If the results are significantly changed by
this commit then I claim the result was not robust.

- Example

With

>>> from sherpa.astro.ui import *
>>> load_pha('3c120_pha2.gz')
>>> load_arf(10, '3c120_meg_1.arf.gz')
>>> load_rmf(10, '3c120_meg_1.rmf.gz')
>>> group_width(10, 2)
>>> set_analysis('wave')
>>> notice(21.4, 21.7)

then prior to this change (well, with CIAO 4.13) we get

>>> get_data_plot(10).xlo
array([21.71, 21.7 , 21.69, 21.68, 21.67, 21.66, 21.65, 21.64, 21.63,
       21.62, 21.61, 21.6 , 21.59, 21.58, 21.57, 21.56, 21.55, 21.54,
       21.53, 21.52, 21.51, 21.5 , 21.49, 21.48, 21.47, 21.46, 21.45,
       21.44, 21.43, 21.42, 21.41, 21.4 ])

>>> get_data_plot(10).xhi
array([21.7 , 21.69, 21.68, 21.67, 21.66, 21.65, 21.64, 21.63, 21.62,
       21.61, 21.6 , 21.59, 21.58, 21.57, 21.56, 21.55, 21.54, 21.53,
       21.52, 21.51, 21.5 , 21.49, 21.48, 21.47, 21.46, 21.45, 21.44,
       21.43, 21.42, 21.41, 21.4 , 21.39])

whereas we now get this without the first or last bins - that is,
after swapping lo/hi orders, the

    21.39 - 21.40 Angstrom
    21.70 - 21.71 Angstrom

bins - which makes sense as we asled for the 21.4-21.7 Angstrom
range.

- Details

1D integrated data sets - so Data1DInt and DataPHA when not using
channels - have been changed to better filter based on the
integrated data ranges. We now treat the high value of a filter
call as exclusive - that is lo <= x < hi - rather than the
previous approach which was sort of lo <= x <= hi but I can't
exactly guarantee that.

You can see that a number of tests are better behaved - e.g.
tests that check when you use elo_i, ehi_i for a filter now
better selects just that bin, and not some version of 1, 2, or
3 bins. The behavior for the first and last bins are now
better behaved too.

Since our limit checks use sao_fcmp (with a 1e-7 tolerance)
rather than straight forward less-than or greater-than checks,
we can end up selecting slightly different bins to XSPEC (e.g.
0.3-0.40000001 gets excluded by our notice(0.4, ...) call but
included by XSPEC) but we are always going to have some
differences when comparing with numerical values.