Remove the _to_channel DataPHA method
It turns out we only used _to_channel in the notice call, and
as we've now changed how we do the filtering we can remove
the _to_chanel call, and the associated channel/energy/wavelength
routines. We had a number of tests of these routines so they
can go away too.