Fix corner case when filtering PHA files
The main change to the code is a clean up of handling of wavelength
limits, to better handle a 0 value: we now no-longer use a "magic"
value (hc / np.finfo(np.float32).tiny ~ 1.05e39) but use the
data to tell us a useful range. However, in doing so I ended up
tweaking some code that determines when we could exit early and
this lead to a corner case improvement when filtering

  notice/ignore(0, 1)  # channel and TLMIN=1
  notice/ignore(1, 1)  # channel and TLMIN=1

It is not clear what we want a notice or ignore call to do when
a wavelength filter is 0.