Internal improvements to parameter setting to catch setup errors
The model class setup contains repeated information which @olaurino
has pointed out is not ideal. One issue is that we create the
parameters as attributes of the object and then send them to
the parent initialiser. This can lead to problems where the
names aren't kept in sync. This commit does not change this, but
instead adds some validation checks

 - when the parameter is added to the model in
   sherpa.models.model.Model.__setattr__ we now ensure that the
   parameter name matches the attribute name

   This check is case insensitive since we have some tests that
   have different cases so I didn't want to change anything

 - for XSPEC models only we check that the parameter names are
   unique in sherpa.astro.xspec.XSModel.__init__

Ideally the second check would be moved to a parent class, but
it's not clear where that should be since it can not be
sherpa.models.model.Model as this would fail for
sherpa.models.model.CompositeModel instances (i.e. I tried this
originally and lots of tests failed). My best guess is
sherpa.models.model.ArithmeticModel but I don't want to make
too-large a change at this time **and** it's the XSPEC model
classes that are being changed rather than other ones (i.e. I
realised the need for this in adding XSPEC 12.12.0 models after
I made an error when adding a class).

Note that I have used assert rather than raising an explicit
error since these failures should only happen when writing or
changing a model and not in user-facing code.