Remove register keyword from C++ code
clang++ is reporting that register is going away, so we may as well
clean it up now rather than later. I think the compilers have
long since moved away from the idea of the user noting the hot path
of a routine, so I don't expect this to make the code significantly
slower.