Address type confusion
This fixes clang++ warnings (implicit conversion to bool). It
doesn't affect the behavior of the code (assuming that NULL is
mapped to false; if it does change things then it would have been
returning true which would have been wrong, but these are in
error conditions so not ones we can test).