Support XSPEC 12.9.1-era functionality
A number of routines have been added to sherpa.astro.xspec._xspec
and some of them are also made available to sherpa.astro.xspec
and sherpa.astro.ui.

- get_xselements

Returns the supported elements and their Z value (it is unlikely
that XSPEC will change the set of elements any time soon but
this lets us pick any changes they may make).

- get_xsabundancies

Return the current set of abundancies

- set_xsabund

Can now be sent a sequence of 30 numbers to set all the values
in one go.

- get_xsatomdb / set_xsatomdb

The ATOMDB version string.

The following low level functions are left for experimentation
and future work (i.e. these are only provided by
sherpa.astro.xspec._xspec):

- get_xspath_abundance / set_xspath_abundance

The path to the "abundances" file (from which the individual
abundance tables are read): is it meaningful for us to query
or change it?

- get_xsabundance_file / set_xsabudance_file

This is the file name for the "abundances" file.

- get_xsxflt_number / get_xsxflt_has_key / get_xsxflt_key / set_xsxflt_key

Access to the FILT keywords for per-spectrum settings. Currently
this is not useful as we do not send the spectrum number to the
model evaluations, but this lets us experiment.

The C++/Python interface code has seen a number of minor
improvements - that PyErr_Format returns a NULL so we can return
with it, and no need to create a stream to store the error message
as we can use PyErr_Format instead.