Avoid potentiao version differences with CI + XSPEC
Change the xspec_version lie we change in the setup_xspec_ds9.sh
script to not care about the version. This lets us update the
minimum supported XSPEC version in setup.cfg without needing to
update this script. I found the need for this by updaing setup.cfg
but not this script, and wondered why CI just failed.